% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutribudgetr.R
\docType{package}
\name{nutribudgetr}
\alias{nutribudgetr}
\alias{nutribudgetr-package}
\title{nutribudgetr}
\description{
An Implementation of SPRING's Nutrition Budget Analysis Tool in R
}
\details{
Despite recognition that creating and sustaining financial support for 
nutrition is critical, nutrition advocates lack access to accurate and timely 
financial data. In response to this challenge, the USAID-funded Strengthening 
Partnerships, Results, and Innovations in Nutrition Globally (SPRING) project 
has developed a tool that will help nutrition stakeholders learn where the 
funding is, understand existing resource gaps, and advocate for adequate 
budgets. This package is an implementation of SPRING's budget analysis tool 
in R.
}
